Refactor project code to TypeScript - Add typescript npm package - Remove cucumber-pretty - Add @cucumber/cucumber npm package - Add @types/cucumber - Add ts-node npm package - Update script with ts-node/register and report format - Update gitignore with cucumber report
